Reordered migration

+++ db/migrate/20160901182121_create_debts.rb
@@ -1,7 +1,7 @@
 class CreateDebts < ActiveRecord::Migration
   def change
     create_table :debts do |t|
-      t.double :amount
+      t.decimal :amount
       t.integer :debtorID
       t.integer :creditorID
       t.boolean :paid
